Actualización: FromGyro trait en sensor.rs
