+++ sim/src/sensor.rs
@@ -2,6 +2,26 @@ use crate::arenito::*;
 use bevy::prelude::*;
 use rand::{prelude::thread_rng, Rng};
 
+/// This trait aims to unify the calculation of a direction vector from
+/// the output of MPU6050's gyroscope.
+/// Tailored specifically for this simulator's application it's assumed
+/// that the X+ axis points (initially) forwards and it controls the roll
+/// of Arenito.
+///
+/// To determine the direction only the yaw and the pitch are considered.
+/// That means that the y (yaw) and z (pitch) components of the gyro values
+/// are used.
+pub trait FromGyro {
+    fn from_gyro(gyro: &Vec3) -> Vec3;
+}
+
+impl FromGyro for Vec3 {
+    /// Creates a unit vector from a rotation vector.
+    fn from_gyro(gyro: &Vec3) -> Vec3 {
+        Vec3::new(gyro.y.cos(), gyro.z.sin(), gyro.y.sin())
+    }
+}
+
 /// This struct's purpose is to generalize "Error Vector" generation.
 /// Since this is a needed step on simulating sensors.
 struct SensorError;
